Reestructuración del API
-Separación de responsabilidades
-Eliminación de métodos que no se usarán en el API de momento
-Documentación interna
-Se agrega paquete para mantener el estandar de clean code
