+++ api/urls.py
@@ -1,12 +1,12 @@
 from django.urls import path
 from rest_framework.urlpatterns import format_suffix_patterns
-from api import views
+from api.views import feria_views, producto_views
 
 urlpatterns = [
-    path('productos/', views.ProductoList().as_view()),
-    path('productos/<int:pk>/', views.ProductoDetail().as_view()),
-    path('ferias/', views.FeriaList().as_view()),
-    path('ferias/<slug:pk>/', views.FeriaDetail().as_view()),
+    path('productos/', producto_views.ProductoList().as_view()),
+    path('productos/<int:pk>/', producto_views.ProductoDetail().as_view()),
+    path('ferias/', feria_views.FeriaList().as_view()),
+    path('ferias/<slug:pk>/', feria_views.FeriaDetail().as_view()),
 ]
 
 urlpatterns = format_suffix_patterns(urlpatterns)
